test-framework/requirements.txt
# API Testing dependencies
requests==2.31.0
pydantic==2.4.2
pytest==7.4.0
pytest-xdist==3.3.1
allure-pytest==2.13.2

# UI Testing dependencies
selenium==4.14.0
webdriver-manager==4.0.0
